fix: 移除缓存截图未判断文件夹是否存在

+++ src-tauri/src/screenshot.rs
@@ -35,8 +35,9 @@ pub fn screenshot(app: AppHandle) -> Vec<ScreenshotImage> {
         .path()
         .resolve("temp_screenshot", BaseDirectory::AppData)
         .unwrap();
-    
-    std::fs::remove_dir_all(&temp_screenshot_folder).unwrap();
+    if std::fs::metadata(&temp_screenshot_folder).is_ok() {
+        std::fs::remove_dir_all(&temp_screenshot_folder).unwrap();
+    }
     std::fs::create_dir(&temp_screenshot_folder).unwrap();
 
     let mut files: Vec<ScreenshotImage> = Vec::new();
